+++ docs/index.rst
@@ -18,6 +18,7 @@ Contents
 
   pages/container.rst
   pages/result.rst
+  pages/maybe.rst
   pages/io.rst
   pages/unsafe.rst
   pages/functions.rst
