+++ returns/pipeline.py
@@ -1,0 +1,44 @@
+# -*- coding: utf-8 -*-
+
+from functools import wraps
+from inspect import iscoroutinefunction
+
+from returns.primitives.exceptions import UnwrapFailedError
+
+
+def is_successful(container):
+    """
+    Determins if a container was successful or not.
+
+    We treat container that raise ``UnwrapFailedError`` on ``.unwrap()``
+    not successful.
+    """
+    try:
+        container.unwrap()
+    except UnwrapFailedError:
+        return False
+    else:
+        return True
+
+
+def pipeline(function):  # noqa: C901
+    """
+    Decorator to enable 'do-notation' context.
+
+    Should be used for series of computations that rely on ``.unwrap`` method.
+
+    Supports both async and regular functions.
+    """
+    if iscoroutinefunction(function):
+        async def decorator(*args, **kwargs):
+            try:
+                return await function(*args, **kwargs)
+            except UnwrapFailedError as exc:
+                return exc.halted_container
+    else:
+        def decorator(*args, **kwargs):
+            try:
+                return function(*args, **kwargs)
+            except UnwrapFailedError as exc:
+                return exc.halted_container
+    return wraps(function)(decorator)
