+++ returns/result.py
@@ -136,21 +136,6 @@ def Failure(inner_value):  # noqa: N802
     return _Failure(inner_value)
 
 
-def is_successful(container):
-    """
-    Determins if a container was successful or not.
-
-    We treat container that raise ``UnwrapFailedError`` on ``.unwrap()``
-    not successful.
-    """
-    try:
-        container.unwrap()
-    except UnwrapFailedError:
-        return False
-    else:
-        return True
-
-
 def safe(function):  # noqa: C901
     """
     Decorator to covert exception throwing function to 'Result' container.
@@ -172,27 +157,4 @@ def safe(function):  # noqa: C901
                 return Success(function(*args, **kwargs))
             except Exception as exc:
                 return Failure(exc)
-    return wraps(function)(decorator)
-
-
-def pipeline(function):  # noqa: C901
-    """
-    Decorator to enable 'do-notation' context.
-
-    Should be used for series of computations that rely on ``.unwrap`` method.
-
-    Supports both async and regular functions.
-    """
-    if iscoroutinefunction(function):
-        async def decorator(*args, **kwargs):
-            try:
-                return await function(*args, **kwargs)
-            except UnwrapFailedError as exc:
-                return exc.halted_container
-    else:
-        def decorator(*args, **kwargs):
-            try:
-                return function(*args, **kwargs)
-            except UnwrapFailedError as exc:
-                return exc.halted_container
     return wraps(function)(decorator)
