+++ setup.cfg
@@ -40,7 +40,7 @@ per-file-ignores =
   # Disable some pydocstyle checks for package:
   returns/*.py: D104
   # Disable imports in `__init__.py`:
-  returns/__init__.py: F401, Z412
+  returns/__init__.py: F401, Z410, Z412
   # There are multiple assert's in tests:
   tests/*.py: S101, Z202, Z440
   # Disable some pydocstyle checks globally:
