+++ typesafety/test_maybe_types/maybe_decorator.test
@@ -10,6 +10,19 @@ reveal_type(test)  # E: Revealed type is 'def () -> returns.maybe.Maybe[builtins
 [/CASE]
 
 
+[CASE maybe_decorator_no_params_optional]
+[disable_cache]
+from typing import Optional
+from returns.maybe import maybe
+
+@maybe
+def test() -> Optional[int]:
+    return 1
+
+reveal_type(test)  # E: Revealed type is 'def () -> returns.maybe.Maybe[builtins.int*]'
+[/CASE]
+
+
 [CASE maybe_composition_no_params]
 [disable_cache]
 from returns.maybe import maybe
